chore: partially deprecating old helpers
