+++ src/utils/helpers.ts
@@ -4,7 +4,9 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
+import path from 'path';
 
+import config from '../config';
 import { OP_PUSHDATA1 } from '../opcodes';
 import { DECIMAL_PLACES, DEFAULT_TX_VERSION, CREATE_TOKEN_TX_VERSION } from '../constants';
 import buffer from 'buffer';
@@ -22,7 +24,7 @@ import Address from '../models/address';
 import { hexToBuffer, unpackToInt } from './buffer';
 import { crypto, encoding, Address as bitcoreAddress } from 'bitcore-lib';
 import { clone } from 'lodash';
-import { ParseError, CreateTokenTxInvalid } from '../errors';
+import { ParseError, AddressError, OutputValueError, ConstantNotSet, CreateTokenTxInvalid, MaximumNumberInputsError, MaximumNumberOutputsError } from '../errors';
 import { ErrorMessages } from '../errorMessages';
 
 /**
@@ -32,20 +34,6 @@ import { ErrorMessages } from '../errorMessages';
  */
 
 const helpers = {
-
-  /**
-   * Round float to closest int
-   *
-   * @param {number} n Number to be rounded
-   *
-   * @return {number} Closest integer to n passed
-   *
-   * @memberof Helpers
-   * @inner
-   */
-  roundFloat(n: number): number {
-    return Math.round(n*100)/100
-  },
 
   /**
    * Get the formatted value with decimal places and thousand separators
@@ -565,6 +553,92 @@ const helpers = {
     const addressObj = new Address(address, { network });
     return addressObj.getType();
   },
+
+  /**
+   * Check if error is one of the expected and return the message
+   * Otherwise, throws the unexpected error
+   *
+   * @param {Error} e Error thrown
+   *
+   * @return {string} Error message
+   * @memberof Helpers
+   * @inner
+   */
+  handlePrepareDataError(e: Error): string {
+    if (e instanceof AddressError ||
+        e instanceof OutputValueError ||
+        e instanceof ConstantNotSet ||
+        e instanceof CreateTokenTxInvalid ||
+        e instanceof MaximumNumberOutputsError ||
+        e instanceof MaximumNumberInputsError) {
+      return e.message;
+    } else {
+      // Unhandled error
+      throw e;
+    }
+  },
+
+  /**
+   * Returns a string with the short version of the id of a transaction
+   * Returns {first12Chars}...{last12Chars}
+   *
+   * @param {string} hash Transaction ID to be shortened
+   *
+   * @return {string}
+   * @memberof Helpers
+   * @inner
+   *
+   */
+  getShortHash(hash: string): string {
+    return `${hash.substring(0,12)}...${hash.substring(52,64)}`;
+  },
+
+  /**
+   * Cleans a string for comparison. Remove multiple spaces, and spaces at the beginning and end, and transform to lowercase.
+   *
+   * @param {string} str String to be cleaned
+   *
+   * @return {string} String after clean
+   * @memberof Helpers
+   * @inner
+   *
+   */
+  cleanupString(str: string): string {
+    return str.replace(/\s\s+/g, ' ').trim().toLowerCase();
+  },
+
+    /**
+   * Get the URL to connect to the websocket from the server URL of the wallet
+   *
+   * @param {string|null} url The server url of the wallet
+   *
+   * @return {string} Websocket URL
+   *
+   * @memberof Helpers
+   * @inner
+   */
+    getWSServerURL(url: string|null = null): string {
+      let serverURL: string;
+      if (url === null) {
+        serverURL = config.getServerUrl();
+      } else {
+        serverURL = url;
+      }
+
+      const pieces: string[] = serverURL.split(':');
+      const firstPiece = pieces.splice(0, 1);
+      let protocol = '';
+      if (firstPiece[0].indexOf('s') > -1) {
+        // Has ssl
+        protocol = 'wss';
+      } else {
+        // No ssl
+        protocol = 'ws';
+      }
+      serverURL = path.join(`${pieces.join(':')}`, 'ws/');
+      serverURL = `${protocol}:/${serverURL}`;
+      return serverURL;
+    },
 }
 
 export default helpers;
